+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+- Add config `batch_id_option` method to set default method to resolve batch
+  value. Previously it was `:id` hardcoded value. Now `:id` is a default value.
 - Raname method to add named batch loader from `.batch_loader` to `.batch`
 - Remove `to_json` and `as_json` methods and config options.
   After ruby JSON gem becomes faster in